microblog/requirements.txt
alembic==1.4.2
appdirs==1.4.4
astroid==2.4.1
attrs==19.3.0
Babel==2.8.0
black==19.10b0
blinker==1.4
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
dnspython==1.16.0
dominate==2.5.1
email-validator==1.1.0
Flask==1.1.2
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
guess-language-spirit==0.5.3
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
pathspec==0.8.0
pop==13
pop-config==6
PyJWT==1.7.1
pylint==2.5.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.17
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
visitor==0.1.3
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
